Moved raster::processing::dem to raster::dem.
