+++ src/raster/dem/color_relief.rs
@@ -3,7 +3,7 @@ use std::path::{Path, PathBuf};
 
 use crate::cpl::CslStringList;
 use crate::errors;
-use crate::raster::processing::dem::options::common_dem_options;
+use crate::raster::dem::options::common_dem_options;
 
 use super::options::CommonOptions;
 
@@ -125,7 +125,7 @@ mod tests {
     use crate::assert_near;
     use crate::cpl::CslStringList;
     use crate::errors::Result;
-    use crate::raster::processing::dem::color_relief;
+    use crate::raster::dem::color_relief;
     use crate::raster::StatisticsAll;
     use crate::test_utils::{fixture, InMemoryFixture};
     use crate::Dataset;
