+++ src/raster/dem/roughness.rs
@@ -2,7 +2,7 @@ use std::num::NonZeroUsize;
 
 use crate::cpl::CslStringList;
 use crate::errors;
-use crate::raster::processing::dem::options::common_dem_options;
+use crate::raster::dem::options::common_dem_options;
 
 use super::options::CommonOptions;
 
@@ -35,8 +35,8 @@ mod tests {
     use crate::assert_near;
     use crate::cpl::CslStringList;
     use crate::errors::Result;
-    use crate::raster::processing::dem::roughness;
-    use crate::raster::processing::dem::roughness::RoughnessOptions;
+    use crate::raster::dem::roughness;
+    use crate::raster::dem::roughness::RoughnessOptions;
     use crate::raster::StatisticsAll;
     use crate::test_utils::{fixture, InMemoryFixture};
     use crate::Dataset;
