+++ src/raster/dem/tri.rs
@@ -2,7 +2,7 @@ use std::num::NonZeroUsize;
 
 use crate::cpl::CslStringList;
 use crate::errors;
-use crate::raster::processing::dem::options::common_dem_options;
+use crate::raster::dem::options::common_dem_options;
 
 use super::options::CommonOptions;
 
@@ -80,7 +80,7 @@ impl DemTriAlg {
 mod tests {
     use crate::assert_near;
     use crate::errors::Result;
-    use crate::raster::processing::dem::terrain_ruggedness_index;
+    use crate::raster::dem::terrain_ruggedness_index;
     use crate::raster::StatisticsAll;
     use crate::test_utils::{fixture, InMemoryFixture};
     use crate::Dataset;
