--- src/raster/processing/mod.rs
@@ -1,3 +1,0 @@
-//! GDAL processing routines.
-
-pub mod dem;
